Mostly done refactoring datasets
