+++ deep_qa/data/datasets/dataset.py
@@ -1,28 +1,16 @@
-import codecs
-import itertools
 import logging
+from collections import defaultdict
 from typing import Dict, List
 
 import numpy
 import tqdm
 
 from ...common.util import add_noise_to_dict_values
-from ...common.params import Params
-from ..data_indexer import DataIndexer
-from ..instances.instance import Instance, TextInstance, IndexedInstance
+from .. import Vocabulary
+from .. import Instance
 
 logger = logging.getLogger(__name__)  # pylint: disable=invalid-name
 
-
-def log_label_counts(instances: List[TextInstance]):
-    labels = [(x.label, x) for x in instances]
-    labels.sort(key=lambda x: str(x[0]))
-    label_counts = [(label, len([x for x in group]))
-                    for label, group in itertools.groupby(labels, lambda x: x[0])]
-    label_count_str = str(label_counts)
-    if len(label_count_str) > 100:
-        label_count_str = label_count_str[:100] + '...'
-    logger.info("Finished reading dataset; label counts: %s", label_count_str)
 
 class Dataset:
     """
@@ -40,18 +28,6 @@ class Dataset:
         """
         self.instances = instances
 
-    def merge(self, other: 'Dataset') -> 'Dataset':
-        """
-        Combine two datasets.  If you call try to merge two Datasets of the same subtype, you will
-        end up with a Dataset of the same type (i.e., calling IndexedDataset.merge() with another
-        IndexedDataset will return an IndexedDataset).  If the types differ, this method currently
-        raises an error, because the underlying Instance objects are not currently type compatible.
-        """
-        if type(self) is type(other):
-            return self.__class__(self.instances + other.instances)
-        else:
-            raise RuntimeError("Cannot merge datasets with different types")
-
     def truncate(self, max_instances: int):
         """
         If there are more instances than `max_instances` in this dataset, returns a new dataset
@@ -63,150 +39,120 @@ class Dataset:
         new_instances = [i for i in self.instances]
         return self.__class__(new_instances[:max_instances])
 
+    def index_instances(self, vocab: Vocabulary):
+        """
+        Converts all ``UnindexedFields`` in all ``Instances`` in this ``Dataset`` into
+        ``IndexedFields``.  This modifies the current object, it does not return a new object.
+        """
+        for instance in tqdm.tqdm(self.instances):
+            instance.index_fields(vocab)
 
-class TextDataset(Dataset):
-    """
-    A Dataset of TextInstances, with a few helper methods.
-
-    TextInstances aren't useful for much with Keras until they've been indexed.  So this class just
-    has methods to read in data from a file and convert it into other kinds of Datasets.
-    """
-    def __init__(self, instances: List[TextInstance], params: Params=None):
-
-        if params is not None:
-            params.assert_empty("TextDataset")
-        super(TextDataset, self).__init__(instances)
-
-    def to_indexed_dataset(self, data_indexer: DataIndexer) -> 'IndexedDataset':
-        '''
-        Converts the Dataset into an IndexedDataset, given a DataIndexer.
-        '''
-        indexed_instances = [instance.to_indexed_instance(data_indexer) for instance in tqdm.tqdm(self.instances)]
-        return IndexedDataset(indexed_instances)
-
-    @staticmethod
-    def read_from_file(filename: str, instance_class, params: Params=None):
-        with codecs.open(filename, 'r', 'utf-8') as input_file:
-            lines = [x.strip() for x in tqdm.tqdm(input_file.readlines())]
-            return TextDataset.read_from_lines(lines, instance_class, params)
-
-    @staticmethod
-    def read_from_lines(lines: List[str], instance_class, params: Params=None):
-        instances = [instance_class.read_from_line(x) for x in lines]
-        log_label_counts(instances)
-        return TextDataset(instances, params)
-
-
-class IndexedDataset(Dataset):
-    """
-    A Dataset of IndexedInstances, with some helper methods.
-
-    IndexedInstances have text sequences replaced with lists of word indices, and are thus able to
-    be padded to consistent lengths and converted to training inputs.
-    """
-    def __init__(self, instances: List[IndexedInstance]):
-        super(IndexedDataset, self).__init__(instances)
-
-    def sort_by_padding(self, sorting_keys: List[str], padding_noise: float=0.0):
+    def sort_by_padding(self, sorting_keys: List[(str, str)], padding_noise: float=0.0):
         """
         Sorts the ``Instances`` in this ``Dataset`` by their padding lengths, using the keys in
-        ``sorting_keys`` (in the order in which they are provided).
+        ``sorting_keys`` (in the order in which they are provided).  ``sorting_keys`` is a list of
+        ``(field_name, padding_key)`` tuples.
         """
         instances_with_lengths = []
         for instance in self.instances:
             padding_lengths = instance.get_padding_lengths()
             if padding_noise > 0.0:
-                padding_lengths = add_noise_to_dict_values(padding_lengths, padding_noise)
-            instance_with_lengths = [padding_lengths[key] for key in sorting_keys] + [instance]
+                noisy_lengths = {}
+                for field_name, field_lengths in padding_lengths:
+                    noisy_lengths[field_name] = add_noise_to_dict_values(field_lengths, padding_noise)
+                padding_lengths = noisy_lengths
+            instance_with_lengths = [padding_lengths[field_name][padding_key]
+                                     for (field_name, padding_key) in sorting_keys] + [instance]
             instances_with_lengths.append(instance_with_lengths)
         instances_with_lengths.sort(key=lambda x: x[:-1])
         self.instances = [instance_with_lengths[-1] for instance_with_lengths in instances_with_lengths]
 
     def padding_lengths(self):
         padding_lengths = {}
-        lengths = [instance.get_padding_lengths() for instance in self.instances]
-        if not lengths:
+        all_instance_lengths = [instance.get_padding_lengths() for instance in self.instances]
+        if not all_instance_lengths:
             return padding_lengths
-        for key in lengths[0]:
-            padding_lengths[key] = max(x[key] if key in x else 0 for x in lengths)
+        all_field_lengths = defaultdict(list)
+        for instance_lengths in all_instance_lengths:
+            for field_name, instance_field_lengths in instance_lengths.items():
+                all_field_lengths[field_name].append(instance_field_lengths)
+        for field_name, field_lengths in all_field_lengths.items():
+            for padding_key in field_lengths[0].keys():
+                max_value = max(x[padding_key] if padding_key in x else 0 for x in field_lengths)
+                padding_lengths[field_name][padding_key] = max_value
         return padding_lengths
 
-    def pad_instances(self, padding_lengths: Dict[str, int]=None, verbose: bool=True):
+    def as_arrays(self,
+                  padding_lengths: Dict[str, Dict[str, int]]=None,
+                  verbose: bool=True) -> Dict[str, List[numpy.array]]:
         """
-        Makes all of the ``IndexedInstances`` in the dataset have the same length by padding them.
-        This ``Dataset`` object doesn't know what things there are in the ``Instance`` to pad, but
-        the ``Instances`` do, and so does the model that called us, passing in a
-        ``padding_lengths`` dictionary.  The keys in that dictionary must match the lengths that
-        the ``Instance`` knows about.
-
-        Given that, this method does two things: (1) it asks each of the ``Instances`` what their
-        padding lengths are, and takes a max (using :func:`~IndexedDataset.padding_lengths()`).  It
-        then reconciles those values with the ``padding_lengths`` we were passed as an argument to
-        this method, and pads the instances with :func:`IndexedInstance.pad()`.  If
-        ``padding_lengths`` has a particular key specified with a value, that value takes
-        precedence over whatever we computed in our data.  TODO(matt): with dynamic padding, we
-        should probably have this be a max padding length, not a hard setting, but that requires
-        some API changes.
-
-        This method `modifies` the current object, it does not return a new ``IndexedDataset``.
+        This method converts this ``Dataset`` into a set of numpy arrays that can be passed through
+        a model.  In order for the numpy arrays to be valid arrays, all ``Instances`` in this
+        dataset need to be padded to the same lengths wherever padding is necessary, so we do that
+        first, then we combine all of the arrays for each field in each instance into a set of
+        batched arrays for each field.
 
         Parameters
         ----------
-        padding_lengths: Dict[str, int]
-            If a key is present in this dictionary with a non-`None` value, we will pad to that
+        padding_lengths : ``Dict[str, Dict[str, int]]``
+            If a key is present in this dictionary with a non-``None`` value, we will pad to that
             length instead of the length calculated from the data.  This lets you, e.g., set a
-            maximum value for sentence length, or word length, if you want to throw out long
-            sequences.
-        verbose: bool, optional (default=True)
+            maximum value for sentence length if you want to throw out long sequences.
+
+            Entries in this dictionary are keyed first by field name (e.g., "question"), then by
+            padding key (e.g., "num_tokens").
+        verbose : ``bool``, optional (default=``True``)
             Should we output logging information when we're doing this padding?  If the dataset is
             large, this is nice to have, because padding a large dataset could take a long time.
             But if you're doing this inside of a data generator, having all of this output per
             batch is a bit obnoxious.
+
+        Returns
+        -------
+        data_arrays : ``Dict[str, List[numpy.array]]``
+            A dictionary of data arrays, keyed by field name, suitable for passing as input to a
+            model.  This is a `batch` of instances, so, e.g., if the instances have a "question"
+            field and an "answer" field, the "question" fields for all of the instances will be
+            grouped together into a single array, and the "answer" fields for all instances will be
+            similarly grouped in a parallel set of arrays, for batched computation.
         """
+        if padding_lengths is None:
+            padding_lengths = defaultdict(dict)
         # First we need to decide _how much_ to pad.  To do that, we find the max length for all
         # relevant padding decisions from the instances themselves.  Then we check whether we were
-        # given a max length for a particular dimension.  If we were, we use that instead of the
-        # instance-based one.
+        # given a max length for a particular field and padding key.  If we were, we use that
+        # instead of the instance-based one.
         if verbose:
             logger.info("Padding dataset of size %d to lengths %s", len(self.instances), str(padding_lengths))
             logger.info("Getting max lengths from instances")
         instance_padding_lengths = self.padding_lengths()
         if verbose:
             logger.info("Instance max lengths: %s", str(instance_padding_lengths))
-        lengths_to_use = {}
-        for key in instance_padding_lengths:
-            if padding_lengths and padding_lengths[key] is not None:
-                lengths_to_use[key] = padding_lengths[key]
-            else:
-                lengths_to_use[key] = instance_padding_lengths[key]
+        lengths_to_use = defaultdict(dict)
+        for field_name, instance_field_lengths in instance_padding_lengths.items():
+            for padding_key in instance_field_lengths:
+                if padding_lengths[field_name].get(padding_key) is not None:
+                    lengths_to_use[field_name][padding_key] = padding_lengths[padding_key]
+                else:
+                    lengths_to_use[field_name][padding_key] = instance_field_lengths[padding_key]
+
+        # Now we actually pad the instances to numpy arrays.
+        field_arrays = defaultdict(list)
         if verbose:
             logger.info("Now actually padding instances to length: %s", str(lengths_to_use))
             for instance in tqdm.tqdm(self.instances):
-                instance.pad(lengths_to_use)
+                for field, arrays in instance.pad(lengths_to_use):
+                    field_arrays[field].append(arrays)
         else:
             for instance in self.instances:
-                instance.pad(lengths_to_use)
+                for field, arrays in instance.pad(lengths_to_use):
+                    field_arrays[field].append(arrays)
 
-    def as_training_data(self):
-        """
-        Takes each ``IndexedInstance`` and converts it into (inputs, labels), according to the
-        Instance's as_training_data() method.  Both the inputs and the labels are numpy arrays.
-        Note that if the ``Instances`` return tuples for their inputs, we convert the list of
-        tuples into a tuple of lists, before converting everything to numpy arrays.
-        """
-        inputs = []
-        labels = []
-        instances = self.instances
-        for instance in instances:
-            instance_inputs, label = instance.as_training_data()
-            inputs.append(instance_inputs)
-            labels.append(label)
-        if isinstance(inputs[0], tuple):
-            inputs = [numpy.asarray(x) for x in zip(*inputs)]
-        else:
-            inputs = numpy.asarray(inputs)
-        if isinstance(labels[0], tuple):
-            labels = [numpy.asarray(x) for x in zip(*labels)]
-        else:
-            labels = numpy.asarray(labels)
-        return inputs, labels
+        # Finally, we combine the arrays that we got for each instance into one big array (or set
+        # of arrays) per field.
+        for field_name, field_array_list in field_arrays.items():
+            if isinstance(field_array_list[0], [list, tuple]):
+                field_arrays[field_name] = [numpy.asarray(x) for x in zip(*field_array_list)]
+            else:
+                field_arrays[field_name] = numpy.asarray(field_array_list)
+        return field_arrays
