+++ deep_qa/data/fields/__init__.py
@@ -1,0 +1,1 @@
+from .field import Field, IndexedField, UnindexedField
